+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -177,6 +177,9 @@ Improvements to clang-tidy
   scripts by adding the `-hide-progress` option to suppress progress and
   informational messages.
 
+- Improved :program:`clang-tidy` by adding the option `RemovedArgs` to remove
+  arguments sent to the compiler when invoking Clang-Tidy.
+
 - Deprecated the :program:`clang-tidy` ``zircon`` module. All checks have been
   moved to the ``fuchsia`` module instead. The ``zircon`` module will be removed
   in the 24th release.